test(mock): add mocks
