+++ pkg/vulnerability/vulnerability.go
@@ -18,6 +18,12 @@ const (
 	DefaultIgnoreFile = ".trivyignore"
 )
 
+type Operation interface {
+	FillInfo(vulns []types.DetectedVulnerability, light bool)
+	Filter(vulns []types.DetectedVulnerability, severities []dbTypes.Severity,
+		ignoreUnfixed bool, ignoreFile string) []types.DetectedVulnerability
+}
+
 type Client struct {
 	dbc db.Operations
 }
